+++ bin/commands/snapshot.py
@@ -20,10 +20,24 @@ def _stash_buffer(quiet):
             time.sleep(0.1)
 
 
-def snapshot(message=None, quiet=False, files=None):
+def _drop_stash_by_message(message):
+    if not message:
+        return
+
+    stashes = subprocess.check_output(['git', 'stash', 'list']).strip().splitlines()
+    for stash in stashes:
+        # include ': ' to make sure the match is the entire message
+        if stash.endswith(': ' + message):
+            stash_ref = stash[:stash.index(':')]
+            subprocess.call(['git', 'stash', 'drop', '--quiet', stash_ref])
+            return
+
+
+def snapshot(message=None, replace=False, quiet=False, files=None):
     """Create a snapshot of the working directory and index.
 
     :param str or unicode message: the message to use when creating the underlying stash
+    :param bool replace: replace any existing snapshot with the same message
     :param bool quiet: suppress all output
     :param list files: a list of pathspecs to specific files to use when creating the snapshot
     """
@@ -35,6 +49,9 @@ def snapshot(message=None, quiet=False, files=None):
     if not status_output:
         messages.info('No local changes to save. No snapshot created.', quiet)
         return
+
+    if replace:
+        _drop_stash_by_message(message)
 
     stash_command = ['git', 'stash', 'push', '--include-untracked']
     stash_command = stash_command if not quiet else stash_command + ['--quiet']