+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This projec
 ### Features
 - **Snapshot**: add result message [#54][]
 - **Changes**: --inverse option to show common commits [#118][] [#134][]
+- **Snapshot**: replace a snapshot with --replace [#143][]
 
 ### Changes
 - **Settings**: remove get command [#135][]
@@ -16,6 +17,7 @@ All notable changes to this project will be documented in this file. This projec
 [#134]: https://github.com/Brickster/git-commands/issues/134
 [#135]: https://github.com/Brickster/git-commands/issues/135
 [#140]: https://github.com/Brickster/git-commands/issues/140
+[#143]: https://github.com/Brickster/git-commands/issues/143
 
 ## [v0.7.0][] - 2019-08-26
 ### Features